+++ source/m4q/m4q.js
@@ -528,7 +528,7 @@ function iif(val1, val2, val3){
 
 // Source: src/core.js
 
-var m4qVersion = "v1.0.0. Built at 18/09/2019 07:05:45";
+var m4qVersion = "v1.0.0. Built at 19/09/2019 10:19:06";
 var regexpSingleTag = /^<([a-z][^\/\0>:\x20\t\r\n\f]*)[\x20\t\r\n\f]*\/?>(?:<\/\1>|)$/i;
 
 var matches = Element.prototype.matches
@@ -1103,8 +1103,11 @@ $.fn.extend({
         }
 
         return this.each(function(){
+            var el = $(this);
             if (typeof this.value !== "undefined") {
                 this.value = value;
+            } else {
+                el.html(value);
             }
         });
     },